🆙update：刪除 button sample 程式碼，分離CVA class 提升程式碼可讀性
