--- libs/src/ui/element/avatar/Button.mdx
@@ -1,60 +1,0 @@
-import { Canvas, Meta, Controls, Story } from '@storybook/blocks';
-import * as ButtonStories from './Button.stories.js';
-
-{/* 👇 Documentation-only page */}
-{/* <Meta title="Documentation" /> */}
-
-{/* 👇 Component documentation page */}
-
-<Meta of={ButtonStories} />
-
-# Button
-
-## 簡介
-
-`Button` 元件執行表單、對話方塊等中的操作，並支援多種尺寸、狀態等。
-
-## 變體
-
-Bootstrap 包括幾個按鈕變體，每個變體都有自己的語義目的，並添加了一些額外的功能以實現更多控制。
-
-{/* <Canvas of={ButtonStories.ButtonIntent} /> */}
-
-> 輔助使用提示：使用顏色新增意義僅提供視覺指示，不會傳達給螢幕閱讀> 器等輔助科技的使用者。請確保內容本身的含義是顯而易見的（例如，具> 有足夠顏色對比度的可見文字），或透過替代方式包含該含義，例如隱藏> 在.visually-hidden 類別中的附加文字。
-
-<Canvas of={ButtonStories.ButtonDefault} />
-
-## Outline 按鈕
-
-需要一個按鈕，但不需要它們帶來的大量背景顏色？將預設修飾符類別替換為.btn-outline-\*刪除任何按鈕上的所有背景圖像和顏色的修飾符類別。
-
-{/* <Canvas of={ButtonStories.ButtonOutline} /> */}
-
-## SIZE
-
-喜歡更大或更小的按鈕嗎？添加.btn-lg 或.btn-sm 以獲得其他尺寸。
-
-{/* <Canvas of={ButtonStories.ButtonSize} /> */}
-
-## 全寬按鈕
-
-結合我們的顯示和間隙實用程序，創建響應式全寬「塊按鈕」堆棧，如 Bootstrap 4 中的按鈕。透過使用實用程式而不是特定於按鈕的類，我們可以更好地控制間距、對齊和回應行為。
-
-{/* <Canvas of={ButtonStories.ButtonFullWidth} /> */}
-
-## icon 按鈕
-
-按鈕的圖示由 preicon, appenicon 屬性指定與配置。讓使用者透過 [Icon Library](https://icones.js.org/collection/ic) 取得 icon
-
-{/* <Canvas of={ButtonStories.ButtonPreIcon} /> */}
-{/* <Canvas of={ButtonStories.ButtonAppIcon} /> */}
-{/* <Canvas of={ButtonStories.ButtonIcon} /> */}
-
-{/* 顯示控鍵 */}
-{/* <Controls of={ButtonStories.ButtonDefault} exclude={['style']} /> */}
-
-{/* 單獨顯示樣式 */}
-{/* <Story of={ButtonStories.ButtonDefault} /> */}
-
-{/* 顯示樣式並可複製程式碼 */}
-{/* <Canvas of={ButtonStories.ButtonDefault} /> */}
