--- libs/src/ui/element/button/button.spec.ts
@@ -1,9 +1,0 @@
-import { mount } from '@vue/test-utils';
-import Button from './button.vue';
-
-describe('Button', () => {
-  it('renders properly', () => {
-    const wrapper = mount(Button, {});
-    expect(wrapper.text()).toContain('Welcome to Button');
-  });
-});
