+++ libs/src/ui/element/button/button.vue
@@ -1,33 +1,79 @@
-<script setup lang="ts">
-import { getThemeClass } from './style';
+<script setup>
+import { computed } from 'vue';
+import { useButtonCVAClass } from './useButtonCVAClass';
+import { getSizeClass } from '@/utils/getStyleClass';
+import Icon from '@/ui/element/Icon/Icon.vue';
 
-interface ButtonProps {
-  variant?: string;
-  theme?: string;
-  content?: string;
-}
-
-withDefaults(defineProps<ButtonProps>(), {
-  variant: 'contained',
-  theme: 'primary',
-  content: '送出',
+// 定義 Props
+const props = defineProps({
+  variant: {
+    type: String,
+    default: 'contained',
+    validator: (value) => ['contained', 'outlined', 'text'].includes(value),
+  },
+  themeColor: {
+    type: String,
+    default: 'primary',
+    validator: (value) =>
+      [
+        'primary',
+        'secondary',
+        'tertiary',
+        'success',
+        'warning',
+        'error',
+        'info',
+      ].includes(value),
+  },
+  isDisable: {
+    type: Boolean,
+  },
+  size: {
+    type: String,
+    default: '',
+    validator: (value) => ['small', 'medium', 'large'].includes(value),
+  },
+  prefix: {
+    type: String,
+  },
+  suffix: {
+    type: String,
+  },
+  customClass: {
+    type: String,
+    default: '',
+  },
 });
 
-interface ClickEvent {
-  /** 點擊事件 */
-  (event: 'click', id: string): void;
-}
+// 引入 CVA Class
+const buttonCVAClass = useButtonCVAClass(props);
 
-defineEmits<ClickEvent>();
+// 計算 icon class
+const iconSizeClass = computed(() => {
+  return getSizeClass('icon',props.size );
+})
+
+// 計算包括 CVA Class 與自定義 class 的按鈕
+const finalButtonClass = computed(() => {
+  return [buttonCVAClass.value, props.customClass].filter(Boolean).join(' ');
+});
 </script>
 
 <template>
-  <button
-    :class="`button ${getThemeClass(variant, theme)}`"
-    @click="$emit('click', '點擊事件')"
-  >
-    <slot></slot>
+  <button :class="buttonCVAClass">
+      <template v-if="prefix">
+          <Icon :class="iconSizeClass" :name="props.prefix" ></Icon>
+      </template>
+      <slot></slot>
+      <template v-if="suffix">
+        <Icon :class="iconSizeClass" :name="props.suffix" ></Icon>
+      </template>
   </button>
 </template>
 
-<style></style>
+<style lang="scss" scoped>
+//@import "src/style/_sassloader_test.scss";
+//.example {
+//  background-color: $sassloader_test-color;
+//}
+</style>
