--- libs/src/ui/element/button/index.ts
@@ -1,3 +1,0 @@
-// @index(['./*.tsx', '!./*.stories.tsx'], f => `export * from '${f.path}'`)
-export * from './button.vue';
-// @endindex
