--- libs/src/ui/element/button/style.ts
@@ -1,17 +1,0 @@
-export const getThemeClass = (variant: string, themeColor: string) => {
-  return `button-${variant} button-${variant}-${themeColor}`;
-};
-
-export const getDisableClass = (variant: string) => {
-  if (variant === 'contained') {
-    return 'button-container-disable';
-  }
-
-  if (variant === 'outlined') {
-    return 'button-outline-disable';
-  }
-
-  if (variant === 'text') {
-    return 'button-text-disable';
-  }
-};
